[Refactor] Split pipeline fragment context manager
