+++ be/src/runtime/external_scan_context_mgr.cpp
@@ -44,6 +44,7 @@
 #include "common/thread/thread.h"
 #include "compute_env/result/result_queue_mgr.h"
 #include "exec/pipeline/fragment_context.h"
+#include "exec/pipeline/fragment_context_manager.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/pipeline/query_context_manager.h"
 #include "runtime/exec_env.h"
