+++ be/src/service/internal_service.cpp
@@ -65,6 +65,7 @@
 #include "compute_env/workgroup/work_group.h"
 #include "exec/file_scanner/file_scanner.h"
 #include "exec/pipeline/fragment_context.h"
+#include "exec/pipeline/fragment_context_manager.h"
 #include "exec/pipeline/fragment_executor.h"
 #include "exec/pipeline/lookup_request.h"
 #include "exec/pipeline/primitives/driver_executor.h"
