+++ be/test/exec/pipeline/sink/export_sink_operator_test.cpp
@@ -25,6 +25,7 @@
 #include "compute_env/workgroup/work_group.h"
 #include "compute_env/workgroup/work_group_manager.h"
 #include "exec/pipeline/fragment_context.h"
+#include "exec/pipeline/fragment_context_manager.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/pipeline/query_context_manager.h"
 #include "gen_cpp/RuntimeProfile_types.h"
